+++ aio_overpass/element.py
@@ -29,6 +29,7 @@ __all__ = (
     "Bbox",
     "GeoJsonDict",
     "OverpassDict",
+    "SpatialDict",
 )
 
 GeoJsonDict: TypeAlias = dict[str, Any]
@@ -52,6 +53,23 @@ This tuple can be understood as any of
 """
 
 
+@dataclass(slots=True)
+class SpatialDict:
+    """
+    Mapping of spatial objects with the ``__geo_interface__`` property.
+
+    Objects of this class have the ``__geo_interface__`` property following a protocol
+    [proposed](https://gist.github.com/sgillies/2217756) by Sean Gillies, which can make
+    it easier to use spatial data in other Python software. An example of this is the ``shape()``
+    function that builds Shapely geometries from any object with the ``__geo_interface__`` property.
+
+    Attributes:
+        __geo_interface__: this is the proposed property that contains the spatial data
+    """
+
+    __geo_interface__: dict
+
+
 class Spatial(ABC):
     """
     Base class for (groups of) geospatial objects.
@@ -59,12 +77,6 @@ class Spatial(ABC):
     Classes that represent spatial features extend this class and implement the
     ``geojson`` property. Exporting objects in the GeoJSON format should make it possible
     to integrate them with other tools for visualization, or further analysis.
-
-    Objects of this class have the ``__geo_interface__`` property to follow a protocol
-    [proposed](https://gist.github.com/sgillies/2217756) by Sean Gillies, which can make
-    it easier to use spatial data in other Python software . An example of this is the ``shape()``
-    function that builds Shapely geometries from any object with the ``__geo_interface__`` property.
-
     The ability to re-import the exported GeoJSON structures as ``Spatial`` objects is not
     considered here.
     """
@@ -89,9 +101,15 @@ class Spatial(ABC):
         raise NotImplementedError
 
     @property
-    def __geo_interface__(self) -> GeoJsonDict:
-        """See ``geojson``."""
-        return self.geojson
+    def geo_interfaces(self) -> Iterator[SpatialDict]:
+        """A mapping of this object to ``SpatialDict``s that implement ``__geo_interface__``."""
+        geojson = self.geojson
+        match geojson["type"]:
+            case "FeatureCollection":
+                for feature in geojson["features"]:
+                    yield SpatialDict(__geo_interface__=feature)
+            case _:
+                yield SpatialDict(__geo_interface__=geojson)
 
 
 @dataclass(slots=True)